PUBDEV-5378: Moved introduced logic from FrameUtil.java to TargetEncodingFrameHelper.java. Cleanup unused imports.
