+++ h2o-core/src/main/java/water/util/FrameUtils.java
@@ -8,10 +8,6 @@ import water.*;
 import water.fvec.*;
 import water.parser.ParseDataset;
 import water.parser.ParseSetup;
-import water.rapids.Rapids;
-import water.rapids.Val;
-import water.rapids.ast.prims.advmath.AstKFold;
-import water.rapids.vals.ValFrame;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -952,28 +948,6 @@ public class FrameUtils {
     for (Lockable l : frs) {
       if (l != null) l.delete();
     }
-  }
-
-  /**
-   *
-   * @param frame
-   * @param name name of the fold column
-   * @param nfolds number of folds
-   * @param seed
-   */
-  static public Frame addKFoldColumn(Frame frame, String name, int nfolds, long seed) {
-    Vec foldVec = frame.anyVec().makeZero();
-    frame.add(name, AstKFold.kfoldColumn(foldVec, nfolds, seed == -1 ? new Random().nextLong() : seed));
-    return frame;
-  }
-
-  /**
-   * @return Frame that is registered in DKV
-   */
-  static public Frame register(Frame frame) {
-    frame._key = Key.make();
-    DKV.put(frame);
-    return frame;
   }
 
   /**
